Added rendering pins from the database
Also have popups when the pin is clicked on
To accomodate that, stopped the user from creating a new pin whenever the map is clicked on
Also moved the rate pin and report pin buttons to the popup

A bunch of the original code in Maps.js erroneously used class to set the classname for jsx elements, which is a reserved word in javascript. Changed all instances of class to className.